requirements.txt
nextcord[voice]
requests
python-dotenv